+++ src/sys/keyboard.rs
@@ -1,6 +1,8 @@
-use crate::api;
-use crate::sys::fs::{FileIO, IO};
 use crate::sys;
+
+#[cfg(target_arch = "x86_64")] // TODO: Remove
+use crate::sys::fs::{FileIO, IO};
+
 use crate::sys::x86::int;
 use crate::sys::x86::port::*;
 
@@ -94,6 +96,7 @@ impl KeyboardLayout {
     }
 }
 
+#[cfg(target_arch = "x86_64")] // TODO: Remove
 impl FileIO for KeyboardLayout {
     fn read(&mut self, buf: &mut [u8]) -> Result<usize, ()> {
         int::without_interrupts(|| {
@@ -163,6 +166,7 @@ impl KeyboardBuffer {
     }
 }
 
+#[cfg(target_arch = "x86_64")] // TODO: Remove
 impl FileIO for KeyboardBuffer {
     fn read(&mut self, buf: &mut [u8]) -> Result<usize, ()> {
         int::without_interrupts(||
@@ -222,7 +226,7 @@ fn interrupt_handler() {
                 match key {
                     // Ctrl + Alt + Del
                     DecodedKey::Unicode('\u{7f}') if is_alt && is_ctrl => {
-                        api::power::reboot()
+                        sys::idt::reset();
                     }
 
                     // [Ctrl +] [Shift +] Tab